+++ src/plugins/intel_cpu/src/nodes/executors/acl/acl_eltwise.cpp
@@ -87,8 +87,8 @@ bool AclEltwiseExecutor::supports(const EltwiseConfig& config) {
         srcDescs[argId - ARG_SRC] = desc;
     }
 
-    if (!aclCommonExecutorSupported({srcDescs[0], dstDescs[0]})) {
-        DEBUG_LOG("ACL common preconditions not met");
+    if (!aclSupported({srcDescs[0], dstDescs[0]})) {
+        DEBUG_LOG("ACL common preconditions are not met");
         return false;
     }
