+++ src/plugins/intel_cpu/src/nodes/executors/acl/acl_pooling.hpp
@@ -54,8 +54,8 @@ public:
     [[nodiscard]] bool isSupported(const PoolingAttrs& poolingAttrs,
                                    const std::vector<MemoryDescPtr>& srcDescs,
                                    const std::vector<MemoryDescPtr>& dstDescs) const override {
-        if (!aclCommonExecutorSupported({srcDescs[0], dstDescs[0]})) {
-            DEBUG_LOG("ACL common preconditions not met");
+        if (!aclSupported({srcDescs[0], dstDescs[0]})) {
+            DEBUG_LOG("ACL common preconditions are not met");
             return false;
         }
         auto isSupportedPrecision = [](const ov::element::Type precision) {
