+++ src/plugins/intel_cpu/src/nodes/executors/kleidiai/kleidiai_gathermatmul.cpp
@@ -23,7 +23,7 @@
 #include "nodes/executors/memory_arguments.hpp"
 #include "openvino/core/except.hpp"
 #include "openvino/core/type/element_type.hpp"
-#include "utils/precision_support.h"
+#include "utils/arm_isa_support.h"
 
 namespace ov::intel_cpu {
