+++ src/plugins/intel_cpu/src/utils/precision_support.cpp
@@ -8,7 +8,6 @@
 #    include "cpu/x64/cpu_isa_traits.hpp"
 #endif
 #if defined(OPENVINO_ARCH_ARM) || defined(OPENVINO_ARCH_ARM64)
-#    include "openvino/core/except.hpp"
 #    include "openvino/runtime/system_conf.hpp"
 #endif
 #include "openvino/core/type/element_type.hpp"
@@ -57,21 +56,4 @@ ov::element::Type defaultFloatPrecision() {
     return ov::element::f32;
 }
 
-#if defined(OPENVINO_ARCH_ARM) || defined(OPENVINO_ARCH_ARM64)
-bool hasArmISASupport(ArmISA isa) {
-    switch (isa) {
-    case ArmISA::ASIMD:
-        return true;  // ARMv8-A baseline, always present
-    case ArmISA::SVE:
-        return with_cpu_sve();  // any SVE vector length
-    case ArmISA::DOTPROD:
-        return with_cpu_arm_dotprod();
-    case ArmISA::I8MM:
-        return with_cpu_arm_i8mm();
-    }
-    // An unhandled ArmISA must never be silently reported as supported: that could let an
-    // executor emit instructions the core lacks. Fail loudly so a newly added ISA is wired up.
-    OPENVINO_THROW("hasArmISASupport: unhandled ArmISA value");
-}
-#endif
 }  // namespace ov::intel_cpu
